Extract RGB color space into sub-module
Moving this color space is the most involved because it has both `u8`
and `f64` (aka `Scalar`) variants, and the implementations of some
functions are different for each variant.
